Extracted commands from act.wizard.cpp
