+++ src/engine/ui/cmd_god/forcetime.h
@@ -1,0 +1,17 @@
+/**
+\file forcetime.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 14.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_ENGINE_UI_CMD_GOD_FORCETIME_H_
+#define BYLINS_SRC_ENGINE_UI_CMD_GOD_FORCETIME_H_
+
+class CharData;
+void DoForcetime(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_ENGINE_UI_CMD_GOD_FORCETIME_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
